Implemented the hitbox interaction system
System for hitbox interaction hinges on a script placed on main camera, if moved, several scripts will not be able to find it. Also, relies on players having a Control1 script specifically. Besides these limitations, it should cover every usecase. Runs using a coroutine with WaitForFixedUpdate because that's the only way to run code after a physics update.