Added Death UpHeavy.
Added a rough implementation of grabs. The current system gives the grab complete control of the player, so there's very little pre-written code. Any time you make a grab, you'll have to manage what happens when the player is grabbed manually, with a script attached to the thing doing the grabbing. It might be possible to make this system better, but I don't think that's necessary, so I'm not doing that at the moment.